Moves user role enum to a shared location
Moves the `Role` enum from the users module to a shared location.

This change promotes code reuse and avoids circular dependencies
by making the `Role` enum accessible to other modules without
creating tight couplings.
