--- src/modules/users/enums/user-role.enum.ts
@@ -1,7 +1,0 @@
-
-
-export enum Role {
-    SUPER_ADMIN = 'super_admin',
-    ADMIN_TIENDA = 'admin_tienda',
-    CLIENTE = 'cliente',
-}
